% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeboard.R
\name{addEntry}
\alias{addEntry}
\title{Add entry to a board
Adds an entry to an existing board}
\usage{
addEntry(board = NULL, project, item, priority, status, assignee)
}
\arguments{
\item{board}{an existing board. if NULL, a new one will be created.}

\item{project}{a project name or vector of project names}

\item{item}{an item or vector of items}

\item{priority}{a priority or vector of priorities. Must only contain "Medium", "High", or "Low"}

\item{status}{a status or vector of statuses. Must only contain "Doing", "To do" or "Done"}

\item{assignee}{an assignee or vector of assignees.}
}
\value{
an updated board
}
\description{
Add entry to a board
Adds an entry to an existing board
}
\details{
Users can add rows to a board (or create a new one) by supplying
    vectors containing all entries for each field. All vectors must be have
    the same length, and there is no recycling.
}
\author{
Giuseppe D'Agostino
}
