% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeboard.R
\name{removeEntry}
\alias{removeEntry}
\title{Remove entry from a board
Removes one or more entries from an existing board based on logical indexing}
\usage{
removeEntry(
  board,
  index = NULL,
  project = NULL,
  item = NULL,
  priority = NULL,
  status = NULL,
  assignee = NULL
)
}
\arguments{
\item{board}{an existing board.}

\item{index}{a row index, or vector of row indices. Default is NULL.}

\item{project}{a project name or vector of project names. Default is NULL.}

\item{item}{an item or vector of items. Default is NULL}

\item{priority}{a priority or vector of priorities. Default is NULL.}

\item{status}{a status or vector of statuses. Default is NULL.}

\item{assignee}{an assignee or vector of assignees. Default is NULL.}
}
\value{
an updated board
}
\description{
Remove entry from a board
Removes one or more entries from an existing board based on logical indexing
}
\details{
The function takes vectors of indices and/or any matching field in
    the board, and removes the matching rows. If an user wants to delete all
    rows corresponding to project "submarine", they can do so by specifying
    \code{project = "submarine"} in the arguments. If they want to remove
    more than one project, they can use \code{project = c("submarine",
    "destroyer")} with any other field.
}
\author{
Giuseppe D'Agostino
}
