% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeboard.R
\name{renderBoard}
\alias{renderBoard}
\title{Render a reactable board
Renders a reactable board in a webpage}
\usage{
renderBoard(data)
}
\arguments{
\item{data}{a board created through \code{makeBoard()}}
}
\value{
a reactable page
}
\description{
Render a reactable board
Renders a reactable board in a webpage
}
\author{
Giuseppe D'Agostino
}
