% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeboard.R
\name{updateEntry}
\alias{updateEntry}
\title{Update entry on a board
Updates one or more entries from a board}
\usage{
updateEntry(
  board,
  index = NULL,
  new_project = NULL,
  new_priority = NULL,
  new_status = NULL,
  new_assignee = NULL,
  nj = new_project,
  ns = new_status,
  np = new_priority,
  na = new_assignee
)
}
\arguments{
\item{board}{an existing board.}

\item{index}{a row index, or vector of row indices. Default is NULL.}

\item{new_project}{a new project name. Default is NULL.}

\item{new_priority}{a new priority or vector of priorities. Default is NULL.}

\item{new_status}{a new status or vector of statuses. Default is NULL.}

\item{assignee}{a new assignee or vector of assignees. Default is NULL.}
}
\value{
an updated board
}
\description{
Update entry on a board
Updates one or more entries from a board
}
\details{
The function takes vectors of indices and/or any matching field in
    the board, and updates the matching rows regarding the project, the status
    and/or the priority.
}
\author{
Giuseppe D'Agostino
}
